% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agent.R
\name{consular_agent_configuration}
\alias{consular_agent_configuration}
\title{Show the configuration of the local agent}
\usage{
consular_agent_configuration()
}
\value{
\code{list}
}
\description{
Show the configuration of the local agent
}
\references{
\url{https://www.consul.io/api/agent.html#read-configuration}
}
