% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agent.R
\name{consular_agent_deregister_service}
\alias{consular_agent_deregister_service}
\title{Deregister a new service from the local agent}
\usage{
consular_agent_deregister_service(id)
}
\arguments{
\item{id}{string}
}
\description{
Deregister a new service from the local agent
}
\examples{
\dontrun{
consular_agent_deregister_service('ssh-on-devbox')
}
}
\references{
\url{https://www.consul.io/api/agent/service.html#deregister-service}
}
