% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agent.R
\name{consular_agent_members}
\alias{consular_agent_members}
\title{List consul agent members}
\usage{
consular_agent_members()
}
\value{
\code{data.table} with a member in each row
}
\description{
List consul agent members
}
