% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agent.R
\name{consular_agent_register_service}
\alias{consular_agent_register_service}
\title{Register a new service to the local agent}
\usage{
consular_agent_register_service(payload)
}
\arguments{
\item{payload}{list describing the service}
}
\description{
Register a new service to the local agent
}
\examples{
\dontrun{
consular_agent_register_service(list(
  Name = 'ssh',
  ID = 'ssh-on-devbox',
  Address = 'localhost',
  Port = 22L
))
}
}
\references{
\url{https://www.consul.io/api/agent/service.html#register-service}
}
