% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agent.R
\name{consular_agent_reload}
\alias{consular_agent_reload}
\title{Reloads the configuration of the local agent}
\usage{
consular_agent_reload()
}
\description{
Reloads the configuration of the local agent
}
\references{
\url{https://www.consul.io/api/agent.html#reaload-configuration}
}
