% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agent.R
\name{consular_agent_services}
\alias{consular_agent_services}
\title{List services registered with the local agent}
\usage{
consular_agent_services()
}
\value{
\code{list}
}
\description{
List services registered with the local agent
}
\references{
\url{https://www.consul.io/api/agent/service.html#list-services}
}
