% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agent.R
\name{consular_agent_register_service}
\alias{consular_agent_register_service}
\title{Register a new service to the local agent}
\usage{
consular_agent_register_service(payload)
}
\arguments{
\item{payload}{list describing the service}
}
\description{
Register a new service to the local agent
}
\examples{
\dontrun{
consular_agent_register_service(list(
  Name = 'ssh',
  ID = 'ssh-on-devbox',
  Address = 'localhost',
  Port = 22L
))

## also set a health-check eg for gRPC
port <- 35829
consular_agent_register_service(list(
  Name = 'test-grpc',
  ID = 'test-grpc-on-localhost',
  Address = 'localhost',
  Port = port,
  Check = list(
    id = 'test-grpc-health-check',
    name = 'test-grpc health check',
    grpc = paste('127.0.0.1', port, sep = ':'),
    interval = '5s',
    grpc_use_tls = FALSE,
    service_id = 'test-grpc-on-localhost'
  )
))
}
}
\references{
\url{https://www.consul.io/api/agent/service.html#register-service}
}
